# openapi.yaml
openapi: 3.0.3
info:
  title: Customer Enrollment API
  description: Northstar Bank uses the Customer Enrollment API to accept, review, and manage digital account enrollment applications across retail banking channels. It is consumed by mobile banking, branch-assisted onboarding, contact-center tools, and approved partner onboarding experiences that need a governed way to submit prospective customer information. The API exists in the Northstar Bank portfolio to standardize enrollment workflows, enforce consistent security controls, and make application status visible across internal and partner journeys.
  version: 1.0.0
  contact:
    name: Northstar Bank API Platform Team
    email: api-platform@northstarbank.example
    url: https://developer.northstarbank.example/apis/customer-enrollment
servers:
  - url: http://localhost:3000
    description: Local demo server
security:
  - ApiKeyAuth: []
tags:
  - name: Health
    description: Service health and readiness checks.
  - name: Applications
    description: Customer enrollment application lifecycle operations.
paths:
  /health:
    get:
      tags:
        - Health
      summary: Check API health
      description: Returns the current service health status for platform monitoring and smoke tests.
      operationId: get-health
      responses:
        "200":
          description: Service is healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                healthy:
                  summary: Healthy service response
                  value:
                    status: healthy
                    service: Customer Enrollment API
                    timestamp: "2026-05-12T16:30:00.000Z"
                    correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /applications:
    post:
      tags:
        - Applications
      summary: Create a customer enrollment application
      description: Creates a new retail banking enrollment application in submitted status.
      operationId: create-application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
            examples:
              checkingApplication:
                summary: Checking account enrollment
                value:
                  productType: checking
                  initialDepositAmount: 250
                  customer:
                    firstName: Riley
                    lastName: Morgan
                    dateOfBirth: "1989-04-18"
                    email: riley.morgan@example.com
                    phone: "+1-555-0172"
                    taxIdLast4: "4821"
                    address:
                      line1: 410 Market Street
                      line2: Suite 8
                      city: Denver
                      state: CO
                      postalCode: "80202"
                      country: US
                  consent:
                    acceptedTerms: true
                    marketingOptIn: false
      responses:
        "201":
          description: Application created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
              examples:
                createdApplication:
                  summary: Created application response
                  value:
                    data:
                      id: app_demo_2001
                      status: submitted
                      productType: checking
                      initialDepositAmount: 250
                      riskScore: 42
                      requiredDocuments:
                        - government-id
                        - proof-of-address
                      customer:
                        firstName: Riley
                        lastName: Morgan
                        dateOfBirth: "1989-04-18"
                        email: riley.morgan@example.com
                        phone: "+1-555-0172"
                        taxIdLast4: "4821"
                        address:
                          line1: 410 Market Street
                          line2: Suite 8
                          city: Denver
                          state: CO
                          postalCode: "80202"
                          country: US
                      consent:
                        acceptedTerms: true
                        marketingOptIn: false
                      decisionSummary: Application submitted for identity verification.
                      createdAt: "2026-05-12T16:30:00.000Z"
                      updatedAt: "2026-05-12T16:30:00.000Z"
                    meta:
                      correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
                      timestamp: "2026-05-12T16:30:00.000Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags:
        - Applications
      summary: List customer enrollment applications
      description: Lists enrollment applications with optional status, product type, and pagination filters.
      operationId: list-applications
      parameters:
        - name: status
          in: query
          required: false
          description: Filter applications by lifecycle status.
          schema:
            type: string
            enum:
              - submitted
              - in-review
              - approved
              - rejected
              - cancelled
            example: submitted
        - name: product-type
          in: query
          required: false
          description: Filter applications by retail banking product type.
          schema:
            type: string
            enum:
              - checking
              - savings
              - credit-card
            example: checking
        - name: limit
          in: query
          required: false
          description: Maximum number of records to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
            example: 25
        - name: offset
          in: query
          required: false
          description: Number of records to skip before returning results.
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        "200":
          description: Applications returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationListResponse"
              examples:
                applicationList:
                  summary: List applications response
                  value:
                    data:
                      - id: app_demo_1001
                        status: submitted
                        productType: checking
                        initialDepositAmount: 250
                        riskScore: 38
                        requiredDocuments:
                          - government-id
                          - proof-of-address
                        customer:
                          firstName: Jordan
                          lastName: Ellis
                          dateOfBirth: "1991-08-14"
                          email: jordan.ellis@example.com
                          phone: "+1-555-0101"
                          taxIdLast4: "3141"
                          address:
                            line1: 1200 Walnut Avenue
                            line2: Apt 4B
                            city: Austin
                            state: TX
                            postalCode: "78701"
                            country: US
                        consent:
                          acceptedTerms: true
                          marketingOptIn: true
                        decisionSummary: Application submitted for identity verification.
                        createdAt: "2026-05-01T14:00:00.000Z"
                        updatedAt: "2026-05-01T14:00:00.000Z"
                    meta:
                      correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
                      timestamp: "2026-05-12T16:30:00.000Z"
                      count: 1
                      total: 1
                      limit: 25
                      offset: 0
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /applications/{application-id}:
    get:
      tags:
        - Applications
      summary: Get a customer enrollment application
      description: Retrieves a single enrollment application by ID.
      operationId: get-application-by-id
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: Application returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
              examples:
                application:
                  summary: Application detail response
                  value:
                    data:
                      id: app_demo_1001
                      status: submitted
                      productType: checking
                      initialDepositAmount: 250
                      riskScore: 38
                      requiredDocuments:
                        - government-id
                        - proof-of-address
                      customer:
                        firstName: Jordan
                        lastName: Ellis
                        dateOfBirth: "1991-08-14"
                        email: jordan.ellis@example.com
                        phone: "+1-555-0101"
                        taxIdLast4: "3141"
                        address:
                          line1: 1200 Walnut Avenue
                          line2: Apt 4B
                          city: Austin
                          state: TX
                          postalCode: "78701"
                          country: US
                      consent:
                        acceptedTerms: true
                        marketingOptIn: true
                      decisionSummary: Application submitted for identity verification.
                      createdAt: "2026-05-01T14:00:00.000Z"
                      updatedAt: "2026-05-01T14:00:00.000Z"
                    meta:
                      correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
                      timestamp: "2026-05-12T16:30:00.000Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Applications
      summary: Delete a customer enrollment application
      description: Removes an enrollment application from the demo data set when it is no longer needed for testing.
      operationId: delete-application
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: Application deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteApplicationResponse"
              examples:
                deleted:
                  summary: Deleted application response
                  value:
                    data:
                      id: app_demo_1001
                      deleted: true
                      status: cancelled
                    meta:
                      correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
                      timestamp: "2026-05-12T16:30:00.000Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /applications/{application-id}/status:
    patch:
      tags:
        - Applications
      summary: Update customer enrollment application status
      description: Moves an application through an approved status transition for enrollment review.
      operationId: update-application-status
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationStatusRequest"
            examples:
              approveApplication:
                summary: Approve an application
                value:
                  status: approved
                  reason: Identity verification and funding checks passed.
      responses:
        "200":
          description: Application status updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
              examples:
                approved:
                  summary: Approved application response
                  value:
                    data:
                      id: app_demo_1002
                      status: approved
                      productType: savings
                      initialDepositAmount: 500
                      riskScore: 21
                      requiredDocuments:
                        - government-id
                      customer:
                        firstName: Avery
                        lastName: Chen
                        dateOfBirth: "1985-11-03"
                        email: avery.chen@example.com
                        phone: "+1-555-0102"
                        taxIdLast4: "2718"
                        address:
                          line1: 88 Lakeview Drive
                          line2: Unit 12
                          city: Portland
                          state: OR
                          postalCode: "97205"
                          country: US
                      consent:
                        acceptedTerms: true
                        marketingOptIn: false
                      decisionSummary: Identity verification and funding checks passed.
                      createdAt: "2026-05-02T15:15:00.000Z"
                      updatedAt: "2026-05-12T16:30:00.000Z"
                    meta:
                      correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
                      timestamp: "2026-05-12T16:30:00.000Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    ApplicationId:
      name: application-id
      in: path
      required: true
      description: Unique identifier for an enrollment application.
      schema:
        type: string
        example: app_demo_1001
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missingField:
              summary: Missing required field
              value:
                error:
                  code: VALIDATION_ERROR
                  message: customer.email is required.
                  correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
    Unauthorized:
      description: API key is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidApiKey:
              summary: Invalid API key
              value:
                error:
                  code: UNAUTHORIZED
                  message: A valid x-api-key header is required.
                  correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missingApplication:
              summary: Application not found
              value:
                error:
                  code: APPLICATION_NOT_FOUND
                  message: Application app_demo_unknown was not found.
                  correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
    Conflict:
      description: Request conflicts with current application state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidTransition:
              summary: Invalid status transition
              value:
                error:
                  code: INVALID_STATUS_TRANSITION
                  message: Cannot transition application from approved to submitted.
                  correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
    InternalServerError:
      description: Unexpected service error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unexpected:
              summary: Unexpected error
              value:
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: An unexpected error occurred.
                  correlationId: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - service
        - timestamp
        - correlationId
      properties:
        status:
          type: string
          example: healthy
        service:
          type: string
          example: Customer Enrollment API
        timestamp:
          type: string
          format: date-time
          example: "2026-05-12T16:30:00.000Z"
        correlationId:
          type: string
          example: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
    CreateApplicationRequest:
      type: object
      required:
        - productType
        - initialDepositAmount
        - customer
        - consent
      properties:
        productType:
          type: string
          enum:
            - checking
            - savings
            - credit-card
          example: checking
        initialDepositAmount:
          type: number
          format: float
          minimum: 0
          example: 250
        customer:
          $ref: "#/components/schemas/Customer"
        consent:
          $ref: "#/components/schemas/Consent"
    UpdateApplicationStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - in-review
            - approved
            - rejected
            - cancelled
          example: approved
        reason:
          type: string
          minLength: 3
          maxLength: 240
          example: Identity verification and funding checks passed.
    ApplicationResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: "#/components/schemas/Application"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ApplicationListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          example:
            - id: app_demo_1001
              status: submitted
              productType: checking
              initialDepositAmount: 250
              riskScore: 38
              requiredDocuments:
                - government-id
                - proof-of-address
              customer:
                firstName: Jordan
                lastName: Ellis
                dateOfBirth: "1991-08-14"
                email: jordan.ellis@example.com
                phone: "+1-555-0101"
                taxIdLast4: "3141"
                address:
                  line1: 1200 Walnut Avenue
                  line2: Apt 4B
                  city: Austin
                  state: TX
                  postalCode: "78701"
                  country: US
              consent:
                acceptedTerms: true
                marketingOptIn: true
              decisionSummary: Application submitted for identity verification.
              createdAt: "2026-05-01T14:00:00.000Z"
              updatedAt: "2026-05-01T14:00:00.000Z"
          items:
            $ref: "#/components/schemas/Application"
        meta:
          $ref: "#/components/schemas/ListResponseMeta"
    DeleteApplicationResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: "#/components/schemas/DeleteApplicationResult"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    DeleteApplicationResult:
      type: object
      required:
        - id
        - deleted
        - status
      properties:
        id:
          type: string
          example: app_demo_1001
        deleted:
          type: boolean
          example: true
        status:
          type: string
          example: cancelled
    Application:
      type: object
      required:
        - id
        - status
        - productType
        - initialDepositAmount
        - riskScore
        - requiredDocuments
        - customer
        - consent
        - decisionSummary
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          example: app_demo_1001
        status:
          type: string
          enum:
            - submitted
            - in-review
            - approved
            - rejected
            - cancelled
          example: submitted
        productType:
          type: string
          enum:
            - checking
            - savings
            - credit-card
          example: checking
        initialDepositAmount:
          type: number
          format: float
          example: 250
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
          example: 38
        requiredDocuments:
          type: array
          example:
            - government-id
            - proof-of-address
          items:
            type: string
            example: government-id
        customer:
          $ref: "#/components/schemas/Customer"
        consent:
          $ref: "#/components/schemas/Consent"
        decisionSummary:
          type: string
          example: Application submitted for identity verification.
        createdAt:
          type: string
          format: date-time
          example: "2026-05-01T14:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-01T14:00:00.000Z"
    Customer:
      type: object
      required:
        - firstName
        - lastName
        - dateOfBirth
        - email
        - phone
        - taxIdLast4
        - address
      properties:
        firstName:
          type: string
          example: Jordan
        lastName:
          type: string
          example: Ellis
        dateOfBirth:
          type: string
          format: date
          example: "1991-08-14"
        email:
          type: string
          format: email
          example: jordan.ellis@example.com
        phone:
          type: string
          example: "+1-555-0101"
        taxIdLast4:
          type: string
          pattern: "^[0-9]{4}$"
          example: "3141"
        address:
          $ref: "#/components/schemas/Address"
    Address:
      type: object
      required:
        - line1
        - city
        - state
        - postalCode
        - country
      properties:
        line1:
          type: string
          example: 1200 Walnut Avenue
        line2:
          type: string
          example: Apt 4B
        city:
          type: string
          example: Austin
        state:
          type: string
          minLength: 2
          maxLength: 2
          example: TX
        postalCode:
          type: string
          example: "78701"
        country:
          type: string
          minLength: 2
          maxLength: 2
          example: US
    Consent:
      type: object
      required:
        - acceptedTerms
        - marketingOptIn
      properties:
        acceptedTerms:
          type: boolean
          example: true
        marketingOptIn:
          type: boolean
          example: false
    ResponseMeta:
      type: object
      required:
        - correlationId
        - timestamp
      properties:
        correlationId:
          type: string
          example: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42
        timestamp:
          type: string
          format: date-time
          example: "2026-05-12T16:30:00.000Z"
    ListResponseMeta:
      allOf:
        - $ref: "#/components/schemas/ResponseMeta"
        - type: object
          required:
            - count
            - total
            - limit
            - offset
          properties:
            count:
              type: integer
              example: 1
            total:
              type: integer
              example: 1
            limit:
              type: integer
              example: 25
            offset:
              type: integer
              example: 0
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - correlationId
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: customer.email is required.
            correlationId:
              type: string
              example: corr_demo_2f8c5f50-8b07-49df-a6ef-10f090783a42